+++ src/main/java/org/intellij/sequencer/config/SequenceSettingsComponent.java
@@ -32,6 +32,7 @@ public class SequenceSettingsComponent implements ActionListener {
     private JTable _colorMapTable;
     private JButton _interfaceColor;
     private JCheckBox _showSimplifyCallName;
+    private JCheckBox _showLambdaCall;
 
     private final ExcludeTableModel _excludeTableModel;
     private final ColorMapTableModel _colorMapTableModel;
@@ -99,6 +100,9 @@ public class SequenceSettingsComponent implements ActionListener {
             return true;
         if(sequenceSettingsState.SHOW_SIMPLIFY_CALL_NAME != _showSimplifyCallName.isSelected())
             return true;
+        if(sequenceSettingsState.SHOW_LAMBDA_CALL != _showLambdaCall.isSelected()) {
+            return true;
+        }
         if(sequenceSettingsState.USE_3D_VIEW != _use3dView.isSelected())
             return true;
         if(!Objects.equals(_fontName.getSelectedItem(), sequenceSettingsState.FONT_NAME))
@@ -122,6 +126,7 @@ public class SequenceSettingsComponent implements ActionListener {
         sequenceSettingsState.SHOW_RETURN_ARROWS = _showReturnArrows.isSelected();
         sequenceSettingsState.SHOW_CALL_NUMBERS = _showCallNumbers.isSelected();
         sequenceSettingsState.SHOW_SIMPLIFY_CALL_NAME = _showSimplifyCallName.isSelected();
+        sequenceSettingsState.SHOW_LAMBDA_CALL = _showLambdaCall.isSelected();
         sequenceSettingsState.USE_3D_VIEW = _use3dView.isSelected();
         sequenceSettingsState.FONT_NAME = (String)_fontName.getSelectedItem();
         sequenceSettingsState.FONT_SIZE = Integer.parseInt(((String) Objects.requireNonNull(_fondSize.getSelectedItem())));
@@ -139,6 +144,7 @@ public class SequenceSettingsComponent implements ActionListener {
         _showReturnArrows.setSelected(sequenceSettingsState.SHOW_RETURN_ARROWS);
         _showCallNumbers.setSelected(sequenceSettingsState.SHOW_CALL_NUMBERS);
         _showSimplifyCallName.setSelected(sequenceSettingsState.SHOW_SIMPLIFY_CALL_NAME);
+        _showLambdaCall.setSelected(sequenceSettingsState.SHOW_LAMBDA_CALL);
         _use3dView.setSelected(sequenceSettingsState.USE_3D_VIEW);
         _fontName.setSelectedItem(sequenceSettingsState.FONT_NAME);
         _fondSize.setSelectedItem(String.valueOf(sequenceSettingsState.FONT_SIZE));